--- app/api/Contact/route.js
@@ -1,26 +1,0 @@
-
-import axios from 'axios';
-
-export default async function handler(req, res) {
-    if (req.method === 'POST') {
-        const formData = req.body;
-
-        const scriptURL = '';
-
-        try {
-            const response = await axios.post(scriptURL, formData, {
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-            });
-
-            res.status(200).json(response.data);
-        } catch (error) {
-            console.error('Error submitting form:', error);
-            res.status(500).json({ error: error.message });
-        }
-    } else {
-        res.setHeader('Allow', ['POST']);
-        res.status(405).end(`Method ${req.method} Not Allowed`);
-    }
-}
